+++ service/matching/semaphore/identifier.go
@@ -5,9 +5,9 @@ import "fmt"
 // Identifier names one semaphore bucket. A semaphore of `size` slots is split into
 // ceil(size/bucket_size) buckets.
 //
-// A bucket is the unit of three things at once, and the design depends on all three lining
-// up: exactly one Matching host owns it, it is one partition of semaphore_tokens, and every
-// grant against it is one conditional write to that partition.
+// A bucket is three things at once: one partition of semaphore_tokens, the target of every
+// conditional write that grants a slot in it, and the unit one Matching host serves. The first
+// two are what make a grant correct; the third only keeps this host's free-set useful.
 //
 // The struct is used directly as a map key, so keep every field comparable. String() is
 // for logs and metrics.
@@ -22,13 +22,13 @@ type Identifier struct {
 // construction rather than on its first grant.
 func NewIdentifier(domainID, semaphoreName string, bucket int) (Identifier, error) {
 	if domainID == "" {
-		return Identifier{}, fmt.Errorf("domainID is required")
+		return Identifier{}, fmt.Errorf("%w: domainID is required", ErrInvalidRequest)
 	}
 	if semaphoreName == "" {
-		return Identifier{}, fmt.Errorf("semaphoreName is required")
+		return Identifier{}, fmt.Errorf("%w: semaphoreName is required", ErrInvalidRequest)
 	}
 	if bucket < 0 {
-		return Identifier{}, fmt.Errorf("bucket must not be negative, got %d", bucket)
+		return Identifier{}, fmt.Errorf("%w: bucket must not be negative, got %d", ErrInvalidRequest, bucket)
 	}
 	return Identifier{DomainID: domainID, SemaphoreName: semaphoreName, Bucket: bucket}, nil
 }
